dts: update starting address of sram0 on g1x0x_g3x0x
The base SRAM apperature starts at 0x2000_0000 which is the
ECC enabled region. However, for the g1x0x_g3x0x family, this
region is deprecated due to inconsistent ECC behavior which
can cause NMIs to occur. This change updates the dts for these
devices to start the sram0 at 0x2020_0000, which is the start
of the unchecked region.

Signed-off-by: Dylan Philpot <d-philpot@ti.com>